backend/services/yolo_service/requirements.txt
litserve==0.2.5
torch==2.9.1
ultralytics==8.3.235
pillow==12.0.0
opencv-python==4.12.0.88
python-dotenv==1.2.1
python-logging-loki==0.3.1
requests==2.32.5
python-multipart==0.0.21
